requirements.txt
PyOpenGL==3.1.7
pygame==2.6.1
numpy==2.1.3
